+++ leda/src/Calendar/CalendarHeader.tsx
@@ -8,7 +8,7 @@ import { CalendarHeaderProps } from './types';
 
 export const CalendarHeader = (props: CalendarHeaderProps): React.ReactElement => {
   const {
-    theme, conditions, viewType, viewDate, onClick, ...restProps
+    theme, conditions, viewType, viewDate, onClick, monthNames, ...restProps
   } = props;
   return (
     <Div className={theme.header} {...restProps}>
@@ -22,7 +22,7 @@ export const CalendarHeader = (props: CalendarHeaderProps): React.ReactElement =
         className={getClassNames(theme.title, { [theme.titleDisabled]: conditions.isTitleDisabled })}
         onClick={(ev) => onClick(CALENDAR_CLICK_ACTION.TITLE_CLICK, ev)}
       >
-        {getCalendarTitle(viewDate, viewType)}
+        {getCalendarTitle(viewDate, viewType, monthNames)}
       </Span>
       <Span
         className={getClassNames(theme.nextButton, { [theme.buttonDisabled]: conditions.isNextButtonDisabled })}